improve clarity of freshness-based scheduling code
